feat(engine): implement GPU-driven rendering pipeline and editor workspace
- Implement Vulkan 1.4 GPU-driven pipeline (transform update, instance count, prefix sum, remap, indirect draw)
- Add multi-viewport editor UI with debug output selectors (Single, Split, Quad layouts)
- Implement asset management for Scenes and Environments with dirty state tracking
- Add async file dialogs and import modals for FBX/glTF scenes and HDR environments
- Add deferred GPU resource retirement via ResourceBin for safe frame management

+++ Assets/SceneCompiler/src/Serialization/CompiledSceneBlobWriter.cpp
@@ -2,6 +2,7 @@
 
 #include <Assets/Core/BlobWriter.hpp>
 #include <Assets/Formats/Texture.hpp>
+#include <Assets/TextureCompiler/CompiledTexture.hpp>
 
 #include <span>
 #include <vector>
@@ -57,7 +58,7 @@ TextureSerializationData buildTextureData(const CompiledScene& scene)
         }
 
         globalMipOffset +=
-            static_cast<uint64_t>(texture.mipMetadata.size()) * sizeof(formats::texture::TextureMipMetadata);
+            texture.mipMetadata.size() * sizeof(formats::texture::TextureMipMetadata);
 
         result.textureData.insert(result.textureData.end(), texture.data.begin(), texture.data.end());
